update MostRecentPatch property for tournamentData
